+++ src/common/line-stream-result.js
@@ -1,0 +1,18 @@
+function initializeResultObject()
+{
+	var dataRes = {};
+	
+	dataRes["successful"] = true;
+	dataRes["lineNumber"] = 0;
+	dataRes["entries"] = [];
+	dataRes["messageText"] = "";
+	
+	return dataRes;
+}
+
+
+
+module.exports =
+{
+	initializeObject: initializeResultObject
+};