+++ changelog.md
@@ -1,8 +1,39 @@
 # Changelog
 
+**./src/common/line-stream-result.js**
+* New file - Defines result object for reading input files line by line.
+
+---
+
+**./src/common/value-prep.js**
+* New file - Contains functions for preparing input values.
+	* Cast string value.
+	* Remove excess whitespace.
+	* Neutralize character casing. - Also saves location of capital letters.
+
+---
+
+**./src/common/validation-tasks.js - New functions**
+* writeLineStreamErrorText
+	* Writes error text for line streaming.
+* checkDataLineLengthNumber
+	* Validates the length of a data line.
+* checkDataEntryLengthNumber'
+	* Validates the length of a data line after it has been sanitized.
+	* Entry name and maximum length can be different based on the particular file.
+	* Positive = Valid
+	* Zero = Empty
+	* Negative = Too long
+
+---
+
 **./src/common/number-limits.js**
-* Adjusted 'aboutQuotes' upper limit from 100 to 10
-* Added new ranges:
-	* nameLength (Max: 20)
-	* accentLength (Max: 20)
-	* dataLength (Max: 150)
+* Changed the following properties from ranges to maximum values:
+	* nameLength
+	* accentLength
+	* dataLength
+
+---
+
+**./src/input/read-line-data.js**
+* New file - Used to read and validate input data .txt files